+++ PiSub Project/PiSub-Project/src/main/java/com/mypisubproject/PiSub/Project/repository/WalletRepository.java
@@ -3,6 +3,6 @@ package com.mypisubproject.PiSub.Project.repository;
 import com.mypisubproject.PiSub.Project.model.Wallet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WalletRepository extends JpaRepository<Wallet, Long> {
+public interface WalletRepository extends JpaRepository<Wallet, Integer> {
 
 }
